requirements.txt
alembic==1.8.0
aniso8601==9.0.1
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
functions==0.7.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.6.0
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.17.0
mccabe==0.7.0
networkx==2.8
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
pygame==2.1.2
pylama==8.3.8
pyparsing==3.0.9
pytz==2022.1
random2==1.0.1
requests==2.28.0
six==1.16.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.37
urllib3==1.26.9
validators==0.19.0
Werkzeug==2.1.2
win32-setctime==1.1.0
WTForms==3.0.1
